requirements.txt



uvicorn
sqlalchemy==2.0.19
fastapi==0.100.0
requests==2.27.1
pydantic==2.0.3
psycopg2-binary==2.9.6
starlette==0.27.0
python-dotenv==1.0.0